cleaned up and organized scripts folder
