+++ scripts/src/config.js
@@ -1,0 +1,9 @@
+export default {
+  firebase: {
+    apiKey: 'AIzaSyCyiMJKUK5__AQfnSDfaQmBnVq-dBL73RY',
+    authDomain: 'polyridesweb.firebaseapp.com',
+    databaseURL: 'https://polyridesweb.firebaseio.com',
+    storageBucket: 'polyridesweb.appspot.com',
+    messagingSenderId: '429026541463',
+  },
+}
